+++ CHANGELOG.md
@@ -16,3 +16,4 @@
  - add clock component to convert seconds to hh:mm:ss format
  - display time passed from the beginning in the client
  - remove client from this repository
+ - server send tree data to the client